+++ src/__tests__/users.int.test.ts
@@ -61,6 +61,7 @@ describe('/users', () => {
 			const dataProvider = await createOrUpdateDataProvider({
 				name: 'Test Provider',
 				shortCode: 'testProvider',
+				keycloakOrganizationId: null,
 			});
 			const funder = await createOrUpdateFunder({
 				name: 'Test Funder',