+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.16.4 2024-12-19
+
 ### Changed
 
 - Upgraded to use OpenAPI Specification 3.1.
+- `DataProvider` type now has `keycloakOrganizationId`, the UUID in Keycloak for the related Organization.
 
 ## 0.16.3 2024-12-02